Fix circular import of cfm loss
