+++ tfmpe/estimators/training/directly.py
@@ -11,7 +11,7 @@ from ...preprocessing.tokens import Tokens
 from ...preprocessing.utils import Labeller
 from ...nn.training import fit_memory_efficient
 
-from . import _cfm_loss
+from .loss import cfm_loss as _cfm_loss
 
 
 def fit_directly(
